vk: Refactor instance management header
