+++ rpcs3/Emu/RSX/VK/VKGSRender.h
@@ -4,7 +4,7 @@
 
 #include "vkutils/descriptors.h"
 #include "vkutils/data_heap.h"
-#include "vkutils/instance.hpp"
+#include "vkutils/instance.h"
 #include "vkutils/sync.h"
 #include "vkutils/swapchain.hpp"
