+++ rpcs3/VKGSRender.vcxproj.filters
@@ -73,6 +73,9 @@
       <Filter>upscalers\fsr1</Filter>
     </ClCompile>
     <ClCompile Include="Emu\RSX\VK\VKCommonPipelineLayout.cpp" />
+    <ClCompile Include="Emu\RSX\VK\vkutils\instance.cpp">
+      <Filter>vkutils</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Emu\RSX\VK\VKCommonDecompiler.h" />
@@ -124,7 +127,7 @@
     <ClInclude Include="Emu\RSX\VK\vkutils\device.h">
       <Filter>vkutils</Filter>
     </ClInclude>
-    <ClInclude Include="Emu\RSX\VK\vkutils\instance.hpp">
+    <ClInclude Include="Emu\RSX\VK\vkutils\instance.h">
       <Filter>vkutils</Filter>
     </ClInclude>
     <ClInclude Include="Emu\RSX\VK\vkutils\memory.h">
